+++ Source/MounteaDialogueSystem/Private/Helpers/MounteaDialogueSystemSettings.cpp
@@ -1,0 +1,14 @@
+// All rights reserved Dominik Pavlicek 2023
+
+
+#include "Helpers/MounteaDialogueSystemSettings.h"
+
+#define LOCTEXT_NAMESPACE "MounteaDialogueSystemSettings"
+
+UMounteaDialogueSystemSettings::UMounteaDialogueSystemSettings()
+{
+	CategoryName = TEXT("Mountea Framework");
+	SectionName = TEXT("Mountea Dialogue System");
+}
+
+#undef LOCTEXT_NAMESPACE