+++ Source/MounteaDialogueSystem/Private/Interfaces/MounteaDialogueWBPInterface.cpp
@@ -1,0 +1,6 @@
+// All rights reserved Dominik Pavlicek 2023
+
+
+#include "Interfaces/MounteaDialogueWBPInterface.h"
+
+// Add default functionality here for any IMounteaDialogueWBPInterface functions that are not pure virtual.